feat(mongo, sync): support for mongoDB + sync job
Added support for mongoDB as the database where
the campaign data is there from which it's
loaded into the global campaigns in memory slice
by a sync job which runs periodically every 1min
for now, in the beginning since the global campaign
in memory is empty it loads all the campaign data
into the slice.

Tested via updating local mongo data with updated_at
and sync picking up the freshly updated data.

Used curl to make requests and check for basic sanity
and working of the targeting logic

Changes made:
- Mongo connection and client object
- Sync/poller go routine that runs every 1 min
- Mongo query based on freshly updated data from the
  last sync time
- Mongo db model structs with bson tags to fetch data
- Mutex lock as part of a single struct of the global
  campaign slice so that incoming requests don't read
  partial or malformed data when a read is happening
- Mock data for campaign slice expanded for more values